feat: governor can now change ipfsUri of list
remember to deal with list and editions in the subgraph
if you needed to, which may be the case.
now this creates a problem (and thus may be
reverted):
how to deal with an ongoing request
before requirements change?
obviously, treat as time of submission
but that's a mess. better to just create
new list instead...