feat: removed storedRulings logic
Before, arbitrator needed to rule() and then someone else
executed executeRuling(). Instead, make challenger pay the 20k
for the storage of the disputeId -> disputeSlot mapping-
So you can rule() in a single call.
This has changed a lot of logic related to contributions
and it should be simpler now.
There was an oversight that made contributions not marked
as withdrawn when calling withdrawOneContribution()