feat: separate counters for pendingWithdraws
Instead of having a global pendingWithdraws counter per dispute
have two counters, one for every party that is being contributed
That way you wouldn't have to attempt to withdraw contributions
of a loser party in order to free the disputeSlot.
Before this, it was impossible to free a disputeSlot
if there was a contribution for the losing party.